fix: importing command api causes process exit
